fix: One component per file for SideNav (#442)
* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Changes to refactor SideNav into separate source files

* Remove script file

* Renamed files

* Changed references to new file names

* Updated paths
