+++ src/index.js
@@ -40,7 +40,7 @@ export { default as PanelGrid } from './Panel/PanelGrid';
 export { Popover } from './Popover/Popover';
 export { default as ProductTile } from './Tile/ProductTile';
 export { SearchInput } from './SearchInput/SearchInput';
-export { SideNav, SideNavList, SideNavListItem } from './SideNavigation/SideNavigation';
+export { default as SideNav } from './SideNavigation/SideNav';
 export { Tab } from './Tabs/Tab';
 export { TabGroup } from './Tabs/TabGroup';
 export { Table } from './Table/Table';
